Hazel nguyen

+++ social-media-angular/src/app/components/edit-profile/edit-profile.component.ts
@@ -53,19 +53,27 @@ export class EditProfileComponent implements OnInit {
   }
 
   UpdateUser() {
-    this.user.userId = this.authService.currentUser.userId;
+    //this.user.userId = this.authService.currentUser.userId;
+
     console.log(this.user.userId)
     this.userService.UpdateUser(this.user).subscribe(updateUser => {
       console.log(updateUser);
+      alert('Successfuly Updated User!');
     })
   }
 
   deleteAccount() {
-    this.user.userId = this.authService.currentUser.userId;
-    console.log("user to be deleted: " + this.user.userId)
-    this.userService.DeleteUser(this.user.userId!).subscribe()
-    alert('Successfully Deleted Account');
-    this.router.navigate(["login"])
+    //this.user.userId = this.authService.currentUser.userId;
+
+    let text = "Sorry to see you go, confirm to delete your account.";
+    if (confirm(text) == true) {
+      console.log("user to be deleted: " + this.user.userId)
+      this.userService.DeleteUser(this.user.userId!).subscribe()
+      alert('Successfully Deleted Account');
+      this.router.navigate(["login"])
+    } else {
+      this.router.navigate(["edit-profile"]);
+    }
   }
 
   profilePicture: string;
